Resolve 404 for existing MDN link
In 2014 a link to the MDN documentation was added to README.md.   This link currently results in a 404 "Not Found We're sorry, we couldn't find what you were looking for."  I propose that during the next commit the MDN Documentation link in README.md is changed to https://developer.mozilla.org/Mozilla/Tech/Firefox_Accounts .


It appears that the current locale-adaptive SLUG for this documentation is https://developer.mozilla.org/Mozilla/Tech/Firefox_Accounts and that it is currently available in en-US and in de.  We should make use of the locale-adaptive URL.

Thank you for your consideration.